Remove unnecessary code and file.
